KAFKA-12934: Move some controller classes to the metadata package
Move some controller classes to the metadata package so that they can be
used with broker snapshots. Rename ControllerTestUtils to
RecordTestUtils.
