+++ metadata/src/main/java/org/apache/kafka/metadata/Replicas.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.controller;
+package org.apache.kafka.metadata;
 
 import java.util.ArrayList;
 import java.util.Arrays;
